Make the backend create the database if it doesn't exist. Also clean-up.
In the process removes a bunch of unused or overly complicated code and
moves everything from backend/database/db_docker.py, which was only used
by the tests, into conftest.py.
