--- dockerfiles/Dockerfile.db
@@ -1,5 +1,0 @@
-
-FROM postgres:latest
-ENV POSTGRES_PASSWORD password
-ENV POSTGRES_DB dt_dev
-COPY init.sql /docker-entrypoint-initdb.d/
