--- init.sql
@@ -1,1 +1,0 @@
-CREATE DATABASE dt_dev;
